+++ src/probnum/diffeq/odefiltsmooth/initialize/__init__.py
@@ -1,6 +1,17 @@
 """Initialisation procedures for ODE filters."""
 
-from ._initialize_with_runge_kutta import initialize_odefilter_with_rk
+from ._initialize import InitializationRoutine
 from ._initialize_with_taylormode import initialize_odefilter_with_taylormode
+from ._runge_kutta import RungeKuttaInitialization
 
-__all__ = ["initialize_odefilter_with_rk", "initialize_odefilter_with_taylormode"]
+__all__ = [
+    "InitializationRoutine",
+    "RungeKuttaInitialization",
+    "initialize_odefilter_with_taylormode",
+]
+
+
+# Set correct module paths (for superclasses).
+# Corrects links and module paths in documentation.
+InitializationRoutine.__module__ = "probnum.diffeq.odefiltsmooth.initialize"
+RungeKuttaInitialization.__module__ = "probnum.diffeq.odefiltsmooth.initialize"
